Moved application wide files to src/app
