+++ frontend/src/app/useSessionState.ts
@@ -1,5 +1,7 @@
 import { useEffect, useState } from "react";
 
+import { ApiClient, DEFAULT_CANCEL_REASON } from "@/api/ApiClient";
+import { ApiResult, isSuccess } from "@/api/ApiResult";
 import {
   LOGIN_REQUEST_BODY,
   LOGIN_REQUEST_PATH,
@@ -7,9 +9,6 @@ import {
   LOGOUT_REQUEST_PATH,
   WHOAMI_REQUEST_PATH,
 } from "@/types/generated/openapi";
-
-import { ApiClient, DEFAULT_CANCEL_REASON } from "./ApiClient";
-import { ApiResult, isSuccess } from "./ApiResult";
 
 export interface SessionState {
   user: LoginResponse | null;
